requirements.txt
aniso8601==9.0.1
boto3==1.18.12
botocore==1.21.12
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
Flask==2.0.1
Flask-Cors==3.0.10
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.9
gunicorn==20.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
MarkupSafe==2.0.1
marshmallow==3.13.0
mysql-connector-python==8.0.26
protobuf==3.17.3
psycopg2-binary==2.9.1
PyJWT==2.1.0
pymongo==3.12.0
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
urllib3==1.26.6
Werkzeug==2.0.1
pyjwt
